refactor: move string normalization logic to separate file
